backend/requirements.txt
Django
djangorestframework
django-cors-headers     
python-dotenv           
requests
pandas
numpy
Pillow
tensorflow
tqdm    
torch
openai-clip
torchvision
